# server/openapi.yaml
openapi: 3.0.3
info:
  title: BaudAgain BBS API
  description: |
    REST API for BaudAgain BBS - A modern AI-enhanced bulletin board system.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Authentication endpoints: 10 requests per minute
    - Read operations: 100 requests per 15 minutes
    - Write operations: 30 requests per minute
    - Message posting: 30 messages per hour
    
  version: 1.0.0
  contact:
    name: BaudAgain BBS
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        handle:
          type: string
        realName:
          type: string
        location:
          type: string
        bio:
          type: string
        accessLevel:
          type: integer
          minimum: 0
          maximum: 255
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
        totalCalls:
          type: integer
        totalPosts:
          type: integer
        preferences:
          type: object
          properties:
            terminalType:
              type: string
            screenWidth:
              type: integer
            screenHeight:
              type: integer
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: Get list of all users (admin only)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [handle, createdAt, lastLogin]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Not authenticated
        '403':
          description: Not authorized (requires admin)
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user
      description: Get user profile by ID
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    patch:
      tags:
        - Users
      summary: Update user
      description: Update user profile
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                realName:
                  type: string
                location:
                  type: string
                bio:
                  type: string
                accessLevel:
                  type: integer
                  minimum: 0
                  maximum: 255
                preferences:
                  type: object
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Not authorized
        '404':
          description: User not found
...